+++ apps/mobile/src/services/init-app-services.ts
@@ -12,8 +12,6 @@ export function initAppServices() {
         bitflowApiHost: process.env.EXPO_PUBLIC_BITFLOW_API_HOST ?? '',
         bitflowApiKey: process.env.EXPO_PUBLIC_BITFLOW_API_KEY ?? '',
         bitflowProviderAddress: process.env.EXPO_PUBLIC_BITFLOW_PROVIDER_ADDRESS ?? '',
-        readonlyCallApiHost: process.env.EXPO_PUBLIC_BITFLOW_READONLY_CALL_API_HOST ?? '',
-        readonlyCallApiKey: process.env.EXPO_PUBLIC_BITFLOW_READONLY_CALL_API_KEY ?? '',
         keeperApiHost: process.env.EXPO_PUBLIC_BITFLOW_KEEPER_API_HOST ?? '',
         keeperApiKey: process.env.EXPO_PUBLIC_BITFLOW_KEEPER_API_KEY ?? '',
       },